Update gup.rc
Regarding this issue: https://github.com/notepad-plus-plus/notepad-plus-plus/issues/14668, I think it can be solved by changing line 75 and also, in line 81 I think it is "Cancel" rather than "Never", as it allows to try again from the "Update Notepad++" option inside the "?" symbol.

It is possible that I am not doing it right. My idea is to collaborate, but I lack a lot of knowledge to do it correctly.

I apologise for the possible mistakes.